bot/pyproject.toml
[project]
name = "bot"
version = "1.0.0"
description = "This app is created to suggest users by requests top flavors/brands by brand/flavor of hookah tobacco."
authors = [
    {name = "Vyacheslav Ilchenko",email = "vdilchenko@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.2"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-telegram-bot (>=22.0,<23.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"

